src/requirements.txt
aioftp==0.21.4
bigtree==0.12.5
certifi==2024.8.30
cffi==1.15.1
compose-go==2.29.3
cramjam==2.8.3
dateparser==1.2.0
dbfread==2.0.7
elasticsearch==7.16.2
fastparquet==2024.5.0
fsspec==2024.9.0
humanize==4.10.0
loguru==0.6.0
numpy==1.26.2
packaging==24.1
pandas==2.2.2
pyarrow==17.0.0
pycparser==2.21
pyreaddbc==1.2.0
pysus==0.14.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
six==1.16.0
tqdm==4.64.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
Unidecode==1.3.8
urllib3==1.26.20
urwid==2.6.15
wcwidth==0.2.13
wget==3.2
